picoTTS/wrapper/requirements.txt
py-picotts==0.1.2
httpserver=1.1.0
